refactor: getIsCurrentUser, templateTime을 helpers로 분리
