+++ src/components/FoodParty/FoodPartyDetail/Chat/MessageListitem.tsx
@@ -1,5 +1,6 @@
 import { Avatar, Box, Flex, Text } from '@chakra-ui/react';
 import { Message } from 'types/foodParty';
+import { getIsCurrentUser, templateTime } from 'utils/helpers/chat';
 import { templatePromiseDate } from 'utils/helpers/foodParty';
 
 const MessageListItem = ({
@@ -65,17 +66,3 @@ const MessageListItem = ({
 };
 
 export default MessageListItem;
-
-const getIsCurrentUser = (targetUserId: number, currentUserId: number) =>
-  targetUserId === currentUserId;
-
-const templateTime = (hour: number, minute: number) => {
-  const minuteStartWithZero = String(minute).padStart(2, '0');
-
-  if (hour === 0) return `오전 12:${minuteStartWithZero}`;
-  if (hour === 12) return `오후 12:${minuteStartWithZero}`;
-  if (hour > 12)
-    return `오후 ${String(hour - 12).padStart(2, '0')}:${minuteStartWithZero}`;
-
-  return `오전 ${String(hour).padStart(2, '0')}:${minuteStartWithZero}`;
-};
